item/destroy

+++ app/controllers/items_controller.rb
@@ -12,12 +12,8 @@ class ItemsController < ApplicationController
   end
 
   def destroy
-    if current_user == @item.user
-      if @item.destroy
-        redirect_to root_path, method: :delete
-      else
-        redirect_to item_path(@item)
-      end
+    if current_user == @item.user && @item.destroy 
+      redirect_to root_path, method: :delete
     else
       redirect_to item_path(@item)
     end
